refactor : package 구조 변경 및 컨벤션 일부 수정
