--- src/main/java/lotto/domain/lottoTicket/LottoTicketsGeneratable.java
@@ -1,7 +1,0 @@
-package lotto.domain.lottoTicket;
-
-import lotto.domain.purchase.PurchasingCount;
-
-public interface LottoTicketsGeneratable {
-	LottoTickets generate(PurchasingCount purchasingCount);
-}
