+++ src/main/java/lotto/domain/strategy/LottoTicketsGenerator.java
@@ -1,0 +1,10 @@
+package lotto.domain.strategy;
+
+import lotto.domain.lottoTicket.LottoTickets;
+import lotto.domain.purchase.PurchasingCount;
+
+public interface LottoTicketsGenerator {
+
+	LottoTickets generate(PurchasingCount purchasingCount);
+
+}
