+++ src/main/java/lotto/domain/strategy/ManualLottoTicketsGenerator.java
@@ -1,17 +1,19 @@
-package lotto.domain.lottoTicket;
+package lotto.domain.strategy;
 
 import static java.util.stream.Collectors.*;
 
 import java.util.List;
 import java.util.stream.IntStream;
 
+import lotto.domain.lottoTicket.LottoTicket;
+import lotto.domain.lottoTicket.LottoTickets;
 import lotto.domain.purchase.PurchasingCount;
 
-public class ManualLottoTicketsFactory implements LottoTicketsGeneratable {
+public class ManualLottoTicketsGenerator implements LottoTicketsGenerator {
 
 	private final List<String> inputManualLottoTickets;
 
-	public ManualLottoTicketsFactory(List<String> inputManualLottoTickets) {
+	public ManualLottoTicketsGenerator(List<String> inputManualLottoTickets) {
 		this.inputManualLottoTickets = inputManualLottoTickets;
 	}
