+++ src/test/java/lotto/domain/strategy/AutoLottoTicketsGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottoTicket;
+package lotto.domain.strategy;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -6,15 +6,18 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import lotto.domain.lottoNumber.LottoNumberCache;
+import lotto.domain.lottoTicket.LottoTickets;
 import lotto.domain.purchase.PurchasingCount;
 
-class AutoLottoTicketsFactoryTest {
+class AutoLottoTicketsGeneratorTest {
+
 	@ParameterizedTest
 	@ValueSource(ints = {1, 4, 7, 10})
-	void generate_NumberOfLottoTicket_ReturnLottoTickets(int value) {
-		PurchasingCount purchasingCount = new PurchasingCount(5);
-		LottoTicketsGeneratable lottoTicketsFactory = new AutoLottoTicketsFactory(LottoNumberCache.values());
+	void generate_NumberOfLottoTicket_GenerateLottoTicketsInstance(int value) {
+		PurchasingCount purchasingCount = new PurchasingCount(value);
+		LottoTicketsGenerator lottoTicketsFactory = new AutoLottoTicketsGenerator(LottoNumberCache.values());
 
 		assertThat(lottoTicketsFactory.generate(purchasingCount)).isInstanceOf(LottoTickets.class);
 	}
+
 }
